Commits: 1. Extracted a new abstract class AWorkplaceMissionView from WorkplaceMissionView 2. Improvment of LoginView about alt+f4 to close application 3. Ready for flexible menu of different AppVersions

+++ OperationGuidance_new/Views/LoginView.cs
@@ -55,13 +55,17 @@ namespace OperationGuidance_new.Views {
                     ClickLogin();
                 }
             };
-            _loginForm.FormClosing += (s, e) => e.Cancel = true;
-            _loginForm.AddButton("登录").Click += (s, e) => ClickLogin();
-            _loginForm.AddButton("退出").Click += (s, e) => {
-                if (WidgetUtils.ShowConfirmPopUp("确认退出？")) {
-                    Process.GetCurrentProcess().Kill();
+            _loginForm.FormClosing += (s, e) => {
+                if (!_loginForm.IsDisposed) {
+                    if (WidgetUtils.ShowConfirmPopUp("确认退出？")) {
+                        Process.GetCurrentProcess().Kill();
+                    } else {
+                        e.Cancel = true;
+                    }
                 }
             };
+            _loginForm.AddButton("登录").Click += (s, e) => ClickLogin();
+            _loginForm.AddButton("退出").Click += (s, e) => _loginForm.Close();
 
             _loginForm.PretendToShowToCreateHandlesForChildren();
             _loginForm.ResizeSelf();
